+++ hss_okta_flutter/CHANGELOG.md
@@ -74,3 +74,4 @@
 * Added setDefaultCredential for iOS
 * Added getToken for Android and iOS
 * Fixed Missing error return for native Browser SignIn and Signout streams
+* Fixed Crash on android when okta.properties has bad values